Additions to auto plan 1
I added the encoders to the auto plan, as well as the gyroscope. The
numbers that the if and while statements I put in are arbitrary, and
probably not correct. I created the public variables for the encoders
and the gyroscope. All the tasks that the robot needs to preform in auto
plan 1 should be coded in.