Updated 2017 code to build cleanly with 2018 WPILib
This involved replacement of the CANTalon class with the new WPI_TalonSRX class. The code builds without errors but I have absolutely no idea if it runs!

Warning: There's one entry in teh project file that contains an absolute path from my system. I'll try to figure out how to remove this but, for the meantime, you'll likely need to modify the cscore path on whichever system you try to build this code.